Add base class for units that interact with the ZTF archive
